+++ R/data.R
@@ -185,14 +185,14 @@
 "KHQ5D_data"
 
 
-#' Utility Index weights of the KHQ5D
+#' Value sets of the KHQ5D
 #' @docType data
 #' 
 #' @keywords dataset
 #' 
 #' @description This data set contains the weights to calculate the utility index of the KHQ5D.
 #'
-#' @format This data frame has 30 rows and the following 5 columns:
+#' @format This data frame has 30 rows and the following 2 columns:
 #' \describe{
 #'   \item{UK}{UK KHQ5D weights. \href{https://pubmed.ncbi.nlm.nih.gov/17641139/}{PubMed}}
 #'   \item{Variables}{FullHealth - Best health, usually indicated as 1;