+++ NEWS.md
@@ -1,6 +1,14 @@
 
-# KHQ 0.1.0 (2021-07-15)
+# KHQ 0.2.0 (2021-08-06)
 
+* Adding a new Comulative Frequency Analysis function (KHQ5DFreq) for the health profiles. 
+
+* Adding 'mean.na' option to KHQScore function to handle missing data
+
+* Adding the option to show the vector with the weights of each KHQ5D score in the KHQ5D function
+
+
+# KHQ 0.1.0 (2021-07-15)
 
 * KHQScores function used to generate scores for each dimension of the KHQ